deadman: DEADMAN — the witness the wedge cannot erase, driven from the timer ISR
Metal boot 11 wedged at 117.668 s and the glass was frozen for the last 91
seconds — 43% of the session — with NO DATA for that stretch. The reason is
structural, not luck: every x86 instrument that measures the compositor is
emitted BY the render-service pass, so spread, rtwit, schedx86, dock and
menubar went silent in ONE event, not a cascade. shell.rs already named the
gap: "on x86 there is NO instrument that survives a starved shell ... on
aarch64 the timer-driven train does survive". This is the x86 half of it.

One `[deadman]` line per second from the APIC timer ISR, emitted
UNCONDITIONALLY including when every counter is zero — so SILENCE IS
DISTINGUISHABLE FROM IDLENESS, which is the one reading boot 11 could not
produce.

  [deadman] up=117 hid=42 pmp=998 hq=3 hid_ms=118 comp_ms=91234 \
            gate=2/8891 dec=00200000 dmg=3/3

Safety, which is the property that matters most here: it never takes a lock
the compositor can hold. COMP_GATE is never touched — holder and hold age are
two relaxed loads of gauges wm already keeps. Queue depth is pal's lock-free
conservation law, NOT the blocking `event_queue_depth()`. Damage uses
`TABLE.try_lock()` and reports `?` rather than waiting, never `table()` (a
blocking IRQ-masked spin that would deadlock outright from an ISR).

TWO FIELDS FROM THE BRIEF ARE NOT BUILT AS SPECIFIED, and both say so loudly
rather than being quietly weakened:

  * `hid_in` was to be incremented inside the EHCI IRQ handler. THAT SITE DOES
    NOT EXIST — ehci/mod.rs:19: "No interrupts: no USBINTR write, no IDT
    vector, no MSI." The rMBP's keyboard and trackpad are POLLED. So `hid=` is
    the qTD retirement (earliest honest point, above every decoder and before
    any pal push) and is PAIRED with `pmp=` (poll passes), so `hid=0` can be
    told apart from "nobody looked" instead of silently conflating them.

  * `ser_wait[0..7]` was to count threads blocked on the serialization lock.
    NOTHING EVER BLOCKS ON IT — COMP_GATE is an AtomicBool try-gate; a loser
    DECLINES and returns. A "threads blocked" field would be structurally
    always zero, reading "nobody was waiting" during the exact wedge it exists
    to explain. `dec=` is the honest analogue with the same job: the per-core
    roster of who tried and was turned away.

Knob-gated `deadman` (UNAOS_DEADMAN=1), default OFF — every hook degrades to
an empty inline shim, so an unarmed kernel carries no counter and no string.
Named on both the x86-all and arm-pi type-check legs (the wm.rs hooks are
arch-neutral, so the shim must compile armed on aarch64 too), and stripped
from aarch64 MEDIA so Pi/Jetson images stay byte-identical.

Gate: ./arroyo check green both arches + 12 cfg legs.
  ✅ x86_64 OK
  ✅ aarch64 OK
  ✅ kernel cfg coverage OK (12 legs)
  ✅ userspace x86_64 OK (4 crates)
  ✅ userspace aarch64 OK (5 crates)
  ✅ midden_core tests OK

+++ unaos/crates/kernel/src/deadman.rs
@@ -230,7 +230,7 @@ mod imp {
     /// One composite pass DECLINED by `COMP_GATE`. Charged to the declining core.
     #[inline]
     pub fn note_decline() {
-        let c = crate::arch::percpu::this_cpu().cpu_index.min(CORES - 1);
+        let c = (crate::arch::percpu::this_cpu().cpu_index as usize).min(CORES - 1);
         DEC[c].fetch_add(1, Relaxed);
     }
